Split genomic_position_table.py file into a submodule
